requirements.txt
# Backend FastAPI Requirements - Compatível Python 3.13
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg[binary]